Different approach to exporting internal
