+++ packages/teams-js/src/public/profile.ts
@@ -1,6 +1,5 @@
 import { sendAndHandleSdkError as sendAndHandleError } from '../internal/communication';
 import { ensureInitialized } from '../internal/internalAPIs';
-import * as internalProfile from '../internal/profile';
 import { validateShowProfileRequest } from '../internal/profileUtil';
 import { FrameContexts } from './constants';
 import { ErrorCode } from './interfaces';
@@ -12,14 +11,6 @@ import { runtime } from './runtime';
  * @beta
  */
 export namespace profile {
-  export import Modality = internalProfile.Modality;
-  export import Persona = internalProfile.Persona;
-  // Even though this type is unused in this file, it is referenced by Persona and thus should
-  // be re-exported to ensure it can be used and be documented publicly.
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  export import PersonaIdentifiers = internalProfile.PersonaIdentifiers;
-  export import TriggerType = internalProfile.TriggerType;
-
   /**
    * Opens a profile card at a specified position to show profile information about a persona.
    * @param showProfileRequest The parameters to position the card and identify the target user.
@@ -37,7 +28,7 @@ export namespace profile {
       }
 
       // Convert the app provided parameters to the form suitable for postMessage.
-      const requestInternal: internalProfile.ShowProfileRequestInternal = {
+      const requestInternal: ShowProfileRequestInternal = {
         modality: showProfileRequest.modality,
         persona: showProfileRequest.persona,
         triggerType: showProfileRequest.triggerType,
@@ -51,6 +42,72 @@ export namespace profile {
 
       resolve(sendAndHandleError('profile.showProfile', requestInternal));
     });
+  }
+
+  /**
+   * The type of modalities that are supported when showing a profile.
+   * Can be provided as an optional hint with the request and will be
+   * respected if the hosting M365 application supports it.
+   *
+   * @beta
+   */
+  export type Modality = 'Card' | 'Expanded';
+
+  /**
+   * The type of the profile trigger.
+   *  - MouseHover: The user hovered a target.
+   *  - Press: The target was pressed with either a mouse click or keyboard key press.
+   *  - AppRequest: The show profile request is happening programmatically, without direct user interaction.
+   *
+   * @beta
+   */
+  export type TriggerType = 'MouseHover' | 'Press' | 'AppRequest';
+
+  /**
+   * The set of identifiers that are supported for resolving the persona.
+   *
+   * At least one is required, and if multiple are provided then only the highest
+   * priority one will be used (AadObjectId > Upn > Smtp).
+   *
+   * @beta
+   */
+  export type PersonaIdentifiers = {
+    /**
+     * The object id in Azure Active Directory.
+     *
+     * This id is guaranteed to be unique for an object within a tenant,
+     * and so if provided will lead to a more performant lookup. It can
+     * be resolved via MS Graph (see https://learn.microsoft.com/graph/api/resources/users
+     * for examples).
+     */
+    readonly AadObjectId?: string;
+
+    /**
+     * The primary SMTP address.
+     */
+    readonly Smtp?: string;
+
+    /**
+     * The user principle name.
+     */
+    readonly Upn?: string;
+  };
+
+  /**
+   * The persona to show the profile for.
+   *
+   * @beta
+   */
+  export interface Persona {
+    /**
+     * The set of identifiers that are supported for resolving the persona.
+     */
+    identifiers: PersonaIdentifiers;
+
+    /**
+     * Optional display name override. If not specified the user's display name will be resolved normally.
+     */
+    displayName?: string;
   }
 
   /**
@@ -81,6 +138,11 @@ export namespace profile {
   }
 
   /**
+   * An internal representation of the showProfile parameters suitable for sending via postMessage.
+   * The hub expects to receive an object of this type.
+   */
+
+  /**
    * Checks if the profile capability is supported by the host
    *
    * @returns boolean to represent whether the profile capability is supported
@@ -90,4 +152,21 @@ export namespace profile {
   export function isSupported(): boolean {
     return runtime.supports.profile ? true : false;
   }
+}
+
+/**
+ * Internal representation of a DOMRect suitable for sending via postMessage.
+ */
+export type Rectangle = {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+};
+
+export interface ShowProfileRequestInternal {
+  modality?: profile.Modality;
+  persona: profile.Persona;
+  targetRectangle: Rectangle;
+  triggerType: profile.TriggerType;
 }
